+++ packages/plugin-calls-ui/src/components/Widget.tsx
@@ -1,22 +1,23 @@
 import Icon from '@erxes/ui/src/components/Icon';
-import Tip from '@erxes/ui/src/components/Tip';
-import { __ } from '@erxes/ui/src/utils';
-import React from 'react';
 import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
-import { NotifButton } from '../styles';
-import WidgetPopover from './WidgetPopover';
 import Popover from 'react-bootstrap/Popover';
+import React from 'react';
+import WidgetPopover from './WidgetPopover';
+import { WidgetWrapper } from '../styles';
+import { __ } from '@erxes/ui/src/utils';
 
 type Props = {
   callUserIntegrations: any;
   setConfig: any;
 };
+
 const Widget = (props: Props) => {
   const content = (
     <Popover id="call-popover" className="call-popover">
       <WidgetPopover autoOpenTab="Keyboard" {...props} />
     </Popover>
   );
+
   return (
     <OverlayTrigger
       trigger="click"
@@ -24,11 +25,9 @@ const Widget = (props: Props) => {
       placement="bottom"
       overlay={content}
     >
-      <NotifButton>
-        <Tip text={__('Call')} placement="bottom">
-          <Icon icon="phone" size={26} />
-        </Tip>
-      </NotifButton>
+      <WidgetWrapper>
+        <Icon icon="phone" size={23} />
+      </WidgetWrapper>
     </OverlayTrigger>
   );
 };