+++ packages/plugin-calls-ui/src/containers/SipProvider.tsx
@@ -1,40 +1,34 @@
+import * as moment from 'moment';
+
 import React, { useState } from 'react';
-import SipProvider from '../components/SipProvider';
+import { gql, useMutation, useQuery, useSubscription } from '@apollo/client';
+import { mutations, queries, subscriptions } from '../graphql';
 
-import IncomingCallContainer from './IncomingCall';
-import WidgetContainer from './Widget';
-import { CALL_DIRECTION_INCOMING } from '../lib/enums';
-import { queries, mutations, subscriptions } from '../graphql';
-import { useQuery, gql, useMutation, useSubscription } from '@apollo/client';
-
-import { ModalTrigger } from '@erxes/ui/src/components';
 import { Alert } from '@erxes/ui/src/utils';
+import { CALL_DIRECTION_INCOMING } from '../lib/enums';
 import CallIntegrationForm from '../components/Form';
-import withCurrentUser from '@erxes/ui/src/auth/containers/withCurrentUser';
+import IncomingCallContainer from './IncomingCall';
+import { ModalTrigger } from '@erxes/ui/src/components';
+import SipProvider from '../components/SipProvider';
 import TerminateSessionForm from '../components/TerminateCallForm';
+import WidgetContainer from './Widget';
 import { setLocalStorage } from '../utils';
 
-import * as moment from 'moment';
 import { getSubdomain } from '@erxes/ui/src/utils/core';
+import withCurrentUser from '@erxes/ui/src/auth/containers/withCurrentUser';
 
 const SipProviderContainer = (props) => {
   const [config, setConfig] = useState(
-    JSON.parse(localStorage.getItem('config:call_integrations')),
+    JSON.parse(localStorage.getItem('config:call_integrations') || '{}'),
   );
-  const callInfo = JSON.parse(localStorage.getItem('callInfo'));
+  const callInfo = JSON.parse(localStorage.getItem('callInfo') || '{}');
   const sessionCode = sessionStorage.getItem('sessioncode');
 
   const isConnectCallRequested = JSON.parse(
-    localStorage.getItem('isConnectCallRequested'),
+    localStorage.getItem('isConnectCallRequested') || '{}',
   );
 
   const { data, loading, error } = useQuery(gql(queries.callUserIntegrations));
-  const {
-    data: activeSession,
-    loading: activeSessionLoading,
-    error: activeSessionError,
-    refetch,
-  } = useQuery(gql(queries.activeSession), {});
 
   const [createActiveSession] = useMutation(gql(mutations.addActiveSession));
   const [removeActiveSession] = useMutation(
@@ -43,7 +37,7 @@ const SipProviderContainer = (props) => {
   const [updateHistoryMutation] = useMutation(gql(mutations.callHistoryEdit));
 
   useSubscription(gql(subscriptions.sessionTerminateRequested), {
-    variables: { subdomain: getSubdomain(), userId: props.currentUser._id },
+    variables: { userId: props.currentUser._id },
     onSubscriptionData: () => {
       if (
         !callInfo?.isRegistered ||
@@ -77,9 +71,7 @@ const SipProviderContainer = (props) => {
 
   const createSession = () => {
     createActiveSession()
-      .then(() => {
-        refetch();
-      })
+      .then(() => {})
       .catch((e) => {
         Alert.error(e.message);
       });
@@ -88,8 +80,6 @@ const SipProviderContainer = (props) => {
   const removeSession = () => {
     removeActiveSession()
       .then(() => {
-        refetch();
-
         if (config) {
           setConfig({
             inboxId: config.inboxId,
@@ -136,14 +126,11 @@ const SipProviderContainer = (props) => {
       });
   };
 
-  if (loading || activeSessionLoading) {
+  if (loading) {
     return null;
   }
   if (error) {
     return Alert.error(error.message);
-  }
-  if (activeSessionError) {
-    return Alert.error(activeSessionError.message);
   }
 
   const { callUserIntegrations } = data;
@@ -172,27 +159,29 @@ const SipProviderContainer = (props) => {
       removeActiveSession={removeSession}
     />
   );
+
   if (!config || !config.inboxId) {
     return (
       <ModalTrigger title="Call Config Modal" content={content} isOpen={true} />
     );
   }
-  if (activeSession && activeSession.callsActiveSession) {
-    if (
-      (activeSession.callsActiveSession?.lastLoginDeviceId !== sessionCode ||
-        isConnectCallRequested ||
-        isConnectCallRequested === 'true') &&
-      !callInfo?.isUnRegistered
-    ) {
-      return (
-        <ModalTrigger
-          title="Call Config Modal"
-          content={terminateContent}
-          isOpen={true}
-        />
-      );
-    }
-  }
+
+  // if (activeSession && activeSession.callsActiveSession) {
+  //   if (
+  //     (activeSession.callsActiveSession?.lastLoginDeviceId !== sessionCode ||
+  //       isConnectCallRequested ||
+  //       isConnectCallRequested === "true") &&
+  //     !callInfo?.isUnRegistered
+  //   ) {
+  //     return (
+  //       <ModalTrigger
+  //         title="Call Config Modal"
+  //         content={terminateContent}
+  //         isOpen={true}
+  //       />
+  //     );
+  //   }
+  // }
   if (!config.isAvailable) {
     return (
       <WidgetContainer
@@ -226,9 +215,9 @@ const SipProviderContainer = (props) => {
         urls: 'stun:stun.l.google.com:19302',
       },
       {
-        urls: 'turn:relay1.expressturn.com:3478',
-        username: 'ef9XU6ND3AYQBGG0VB',
-        credential: '7niiKgbs4Kk92V0d',
+        urls: 'turn:relay8.expressturn.com:3478',
+        username: 'efVCM7AV4B0436ZEJQ',
+        credential: 'PtBTQUgzOtZ1T874',
       },
     ],
   };
@@ -237,24 +226,24 @@ const SipProviderContainer = (props) => {
     <SipProvider
       {...sipConfig}
       createSession={createSession}
-      callsActiveSession={activeSession?.callsActiveSession}
       updateHistory={updateHistory}
       callUserIntegration={filteredIntegration}
     >
-      {(state) =>
-        state?.callDirection === CALL_DIRECTION_INCOMING ? (
-          <IncomingCallContainer
-            {...props}
-            callUserIntegrations={callUserIntegrations}
-          />
-        ) : (
+      {(state) => (
+        <>
+          {state?.callDirection === CALL_DIRECTION_INCOMING && (
+            <IncomingCallContainer
+              {...props}
+              callUserIntegrations={callUserIntegrations}
+            />
+          )}
           <WidgetContainer
             {...props}
             callUserIntegrations={callUserIntegrations}
             setConfig={handleSetConfig}
           />
-        )
-      }
+        </>
+      )}
     </SipProvider>
   );
 };